Extracted `ColumnStatsIndexHelper` separating out column-stats index management from Z-ordering
