+++ hudi-client/hudi-spark-client/src/test/java/org/apache/hudi/index/zorder/TestColumnStatsIndexHelper.java
@@ -25,11 +25,11 @@ import java.util.Arrays;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class TestZOrderingIndexHelper {
+public class TestColumnStatsIndexHelper {
 
   @Test
   public void testMergeSql() {
-    String q = ZOrderingIndexHelper.createIndexMergeSql("old", "new", Arrays.asList("file", "a", "b"));
+    String q = ColumnStatsIndexHelper.createIndexMergeSql("old", "new", Arrays.asList("file", "a", "b"));
     assertEquals(
         "SELECT "
             + "if (new.file is null, old.file, new.file) AS file, "
